tun

Trait AbstractDevice

pub trait AbstractDevice: Read + Write {

    // Required methods
    fn tun_index(&self) -> Result<i32>;
    fn tun_name(&self) -> Result<String>;
    fn set_tun_name(&mut self, tun_name: &str) -> Result<()>;
    fn enabled(&mut self, value: bool) -> Result<()>;
    fn address(&self) -> Result<IpAddr>;
    fn set_address(&mut self, value: IpAddr) -> Result<()>;
    fn destination(&self) -> Result<IpAddr>;
    fn set_destination(&mut self, value: IpAddr) -> Result<()>;
    fn broadcast(&self) -> Result<IpAddr>;
    fn set_broadcast(&mut self, value: IpAddr) -> Result<()>;
    fn netmask(&self) -> Result<IpAddr>;
    fn set_netmask(&mut self, value: IpAddr) -> Result<()>;
    fn mtu(&self) -> Result<u16>;
    fn set_mtu(&mut self, value: u16) -> Result<()>;
    fn packet_information(&self) -> bool;

    // Provided method
    fn configure(&mut self, config: &Configuration) -> Result<()> { ... }

}

A TUN abstract device interface.

Required Methods

fn tun_index(&self) -> Result<i32>

Get the device index.

fn tun_name(&self) -> Result<String>

Get the device tun name.

fn set_tun_name(&mut self, tun_name: &str) -> Result<()>

Set the device tun name.

fn enabled(&mut self, value: bool) -> Result<()>

Turn on or off the interface.

fn address(&self) -> Result<IpAddr>

Get the address.

fn set_address(&mut self, value: IpAddr) -> Result<()>

Set the address.

fn destination(&self) -> Result<IpAddr>

Get the destination address.

fn set_destination(&mut self, value: IpAddr) -> Result<()>

Set the destination address.

fn broadcast(&self) -> Result<IpAddr>

Get the broadcast address.

fn set_broadcast(&mut self, value: IpAddr) -> Result<()>

Set the broadcast address.

fn netmask(&self) -> Result<IpAddr>

Get the netmask.

fn set_netmask(&mut self, value: IpAddr) -> Result<()>

Set the netmask.

fn mtu(&self) -> Result<u16>

Get the MTU.

fn set_mtu(&mut self, value: u16) -> Result<()>

Set the MTU.

[Note: This setting has no effect on the Windows platform due to the mtu of wintun is always 65535. –end note]

fn packet_information(&self) -> bool

Return whether the underlying tun device on the platform has packet information

[Note: This value is not used to specify whether the packets delivered from/to tun have packet information. – end note]

Provided Methods

fn configure(&mut self, config: &Configuration) -> Result<()>

Reconfigure the device.